raft: move offset_translator build to storage
This moves the files related to the raft::offset_translator into the
storage build, so the storage layer can own offset translation long
term.

This doesn't yet rename the namespace of the class, just the files. The
namespace change will come in a follow-up change.
