+++ src/v/archival/tests/ntp_archiver_test.cc
@@ -26,9 +26,9 @@
 #include "model/metadata.h"
 #include "net/types.h"
 #include "net/unresolved_address.h"
-#include "raft/offset_translator.h"
 #include "ssx/sformat.h"
 #include "storage/disk_log_impl.h"
+#include "storage/offset_translator.h"
 #include "storage/parser.h"
 #include "storage/storage_resources.h"
 #include "storage/tests/utils/disk_log_builder.h"
