+++ src/v/cluster/controller_backend.cc
@@ -34,6 +34,7 @@
 #include "raft/types.h"
 #include "ssx/event.h"
 #include "ssx/future-util.h"
+#include "storage/offset_translator.h"
 
 #include <seastar/core/abort_source.hh>
 #include <seastar/core/coroutine.hh>
