+++ src/v/cluster/partition_manager.cc
@@ -27,11 +27,11 @@
 #include "raft/consensus.h"
 #include "raft/consensus_utils.h"
 #include "raft/group_configuration.h"
-#include "raft/offset_translator.h"
 #include "raft/rpc_client_protocol.h"
 #include "raft/types.h"
 #include "resource_mgmt/io_priority.h"
 #include "ssx/async-clear.h"
+#include "storage/offset_translator.h"
 #include "storage/offset_translator_state.h"
 #include "storage/segment_utils.h"
 #include "storage/snapshot.h"
