+++ src/v/raft/CMakeLists.txt
@@ -30,7 +30,6 @@ v_cc_library(
     group_configuration.cc
     append_entries_buffer.cc
     follower_queue.cc
-    offset_translator.cc
     recovery_memory_quota.cc
     coordinated_recovery_throttle.cc
     heartbeats.cc
