+++ src/v/raft/consensus_utils.cc
@@ -20,7 +20,6 @@
 #include "model/timestamp.h"
 #include "raft/group_configuration.h"
 #include "raft/logger.h"
-#include "raft/offset_translator.h"
 #include "raft/types.h"
 #include "random/generators.h"
 #include "reflection/adl.h"
@@ -31,6 +30,7 @@
 #include "storage/fs_utils.h"
 #include "storage/kvstore.h"
 #include "storage/ntp_config.h"
+#include "storage/offset_translator.h"
 #include "storage/offset_translator_state.h"
 #include "storage/record_batch_builder.h"
 #include "storage/segment_appender_utils.h"
