+++ src/v/storage/offset_translator.h
